1. 데미지 상태 추가. 여기서 회피, 방어 등 데미지 계산
2. 스테이터스 관련 구조체 추가

3. 공격 딜레이 추가?
- 예를 들면 몬스터가 준비동작(2초) 끝나고 공격 이런 것들을 가능하게끔 딜레이 타임을 받을 수 있게함

4. 몬스터 애니메이션 추가

5. 몬스터 AI 작업 중

+++ Assets/01.Scripts/System/MessageParam.cs
@@ -7,5 +7,5 @@ public class MessageParam
 	public MapObject sender;
 	public MapObject receiver;
 	public string message;
-	public int attackPoint;
+	public float attackPoint;
 }
